requirements.txt
oidn-python
numpy